n_mag reduced to gehrels_err
