--- diffhtwo/experimental/tests/test_n_mag.py
@@ -1,259 +1,0 @@
-import os
-from pathlib import Path
-
-import jax.numpy as jnp
-import numpy as np
-from diffsky.experimental import lc_phot_kern
-from diffsky.experimental import mc_lightcone_halos as mclh
-from diffsky.experimental import precompute_ssp_phot as psspp
-from diffsky.experimental.scatter import DEFAULT_SCATTER_PARAMS
-from diffsky.param_utils.spspop_param_utils import (
-    DEFAULT_SPSPOP_PARAMS,
-)
-from diffsky.ssp_err_model.defaults import ZERO_SSPERR_PARAMS
-from diffstar.defaults import FB, T_TABLE_MIN
-from diffstar.diffstarpop.defaults import (
-    DEFAULT_DIFFSTARPOP_PARAMS,
-)
-from diffstar.diffstarpop.kernels.params.params_diffstarpopfits_mgash import (
-    DiffstarPop_Params_Diffstarpopfits_mgash,
-)
-from dsps.cosmology import flat_wcdm
-from dsps.cosmology.defaults import DEFAULT_COSMOLOGY
-from dsps.data_loaders import retrieve_fake_fsps_data
-from dsps.metallicity.umzr import DEFAULT_MZR_PARAMS
-from jax import random as jran
-from jax.flatten_util import ravel_pytree
-
-from .. import n_mag
-from ..data_loaders import retrieve_tcurves
-from ..lc_utils import zbin_volume
-
-BASE_PATH = Path(__file__).resolve().parent.parent
-LH_CENTROIDS_PATH = BASE_PATH / "data/lh_centroids"
-
-ssp_data = retrieve_fake_fsps_data.load_fake_ssp_data()
-
-DIFFSTARPOP_UM_plus_exsitu = DiffstarPop_Params_Diffstarpopfits_mgash["smdpl_dr1"]
-
-u_diffstarpop_theta_default, u_diffstarpop_unravel = ravel_pytree(
-    DEFAULT_DIFFSTARPOP_PARAMS
-)
-
-
-def test_n_mag():
-    zbins = np.array(
-        [
-            [0.2, 0.5],
-            [1.5, 1.75],
-        ]
-    )
-
-    # Halo lightcone
-    ran_key = jran.key(0)
-    lc_halopop_multi_z = []
-    for zbin in range(0, len(zbins)):
-        ran_key, lc_key = jran.split(ran_key, 2)
-        lgmp_min = 10.0
-        sky_area_degsq = 0.1
-        lc_vol = zbin_volume(
-            sky_area_degsq, zlow=zbins[zbin][0], zhigh=zbins[zbin][1]
-        ).value
-        lc_vol = jnp.array(lc_vol)
-
-        """weighted mc lightcone"""
-        num_halos = 5000
-        lgmp_max = 15.0
-        args = (
-            lc_key,
-            num_halos,
-            zbins[zbin][0],
-            zbins[zbin][1],
-            lgmp_min,
-            lgmp_max,
-            sky_area_degsq,
-        )
-        lc_halopop = mclh.mc_weighted_halo_lightcone(*args)
-        lc_halopop["lc_vol_Mpc3"] = lc_vol
-        lc_halopop_multi_z.append(lc_halopop)
-
-    # Transmission curves
-    tcurves = [
-        retrieve_tcurves.MegaCam_uS,
-        retrieve_tcurves.HSC_G,
-        retrieve_tcurves.HSC_R,
-        retrieve_tcurves.HSC_I,
-        retrieve_tcurves.HSC_Z,
-    ]
-
-    mag_columns = [3]
-    mag_thresh_column = 3
-    mag_thresh = 24.5
-    dmag = 0.2
-    frac_cat = 1.0
-
-    n_z_phot_table = 15
-
-    lc_halopop_z_obs_multi_z = []
-    lc_halopop_t_obs_multi_z = []
-    lc_halopop_mah_params_multi_z = []
-    lc_halopop_nhalos_multi_z = []
-    lc_halopop_logmp0_multi_z = []
-    lc_halopop_vol_mpc3_multi_z = []
-
-    precomputed_ssp_mag_table_multi_z = []
-    z_phot_table_multi_z = []
-    wave_eff_table_multi_z = []
-
-    lh_centroids_multi_z = []
-    dmag_centroids_multi_z = []
-
-    t_0 = flat_wcdm.age_at_z0(*DEFAULT_COSMOLOGY)
-    lgt0 = jnp.log10(t_0)
-    t_table = jnp.linspace(T_TABLE_MIN, 10**lgt0, 100)
-
-    for zbin in range(0, len(zbins)):
-        zmin = zbins[zbin][0]
-        zmax = zbins[zbin][1]
-
-        z_phot_table = jnp.linspace(zmin, zmax, n_z_phot_table)
-
-        precomputed_ssp_mag_table = psspp.get_precompute_ssp_mag_redshift_table(
-            tcurves, ssp_data, z_phot_table, DEFAULT_COSMOLOGY
-        )
-
-        wave_eff_table = lc_phot_kern.get_wave_eff_table(z_phot_table, tcurves)
-
-        lc_halopop = lc_halopop_multi_z[zbin]
-
-        lh_centroids = jnp.asarray(
-            np.load(
-                os.path.join(
-                    LH_CENTROIDS_PATH,
-                    "feniks_lh_centroids_z_"
-                    + str(zbins[zbin][0])
-                    + "-"
-                    + str(zbins[zbin][1])
-                    + "_test.npy",
-                )
-            )
-        )
-        dmag_centroids = jnp.ones((lh_centroids.shape[0], 1)) * dmag
-
-        lc_halopop_z_obs_multi_z.append(lc_halopop["z_obs"])
-        lc_halopop_t_obs_multi_z.append(lc_halopop["t_obs"])
-        lc_halopop_mah_params_multi_z.append(lc_halopop["mah_params"])
-        lc_halopop_logmp0_multi_z.append(lc_halopop["logmp0"])
-        lc_halopop_nhalos_multi_z.append(lc_halopop["nhalos"])
-        lc_halopop_vol_mpc3_multi_z.append(lc_halopop["lc_vol_Mpc3"])
-        precomputed_ssp_mag_table_multi_z.append(precomputed_ssp_mag_table)
-        z_phot_table_multi_z.append(z_phot_table)
-        wave_eff_table_multi_z.append(wave_eff_table)
-        lh_centroids_multi_z.append(lh_centroids)
-        dmag_centroids_multi_z.append(dmag_centroids)
-
-    lc_halopop_z_obs_multi_z = jnp.asarray(lc_halopop_z_obs_multi_z)
-    lc_halopop_t_obs_multi_z = jnp.asarray(lc_halopop_t_obs_multi_z)
-    lc_halopop_mah_params_multi_z = jnp.asarray(lc_halopop_mah_params_multi_z)
-    lc_halopop_logmp0_multi_z = jnp.asarray(lc_halopop_logmp0_multi_z)
-    lc_halopop_nhalos_multi_z = jnp.asarray(lc_halopop_nhalos_multi_z)
-    lc_halopop_vol_mpc3_multi_z = jnp.asarray(lc_halopop_vol_mpc3_multi_z)
-    precomputed_ssp_mag_table_multi_z = jnp.asarray(precomputed_ssp_mag_table_multi_z)
-    z_phot_table_multi_z = jnp.asarray(z_phot_table_multi_z)
-    wave_eff_table_multi_z = jnp.asarray(wave_eff_table_multi_z)
-    lh_centroids_multi_z = jnp.asarray(lh_centroids_multi_z)
-    dmag_centroids_multi_z = jnp.asarray(dmag_centroids_multi_z)
-
-    ran_key, n_key = jran.split(ran_key, 2)
-    n_args_multi_z = (
-        DEFAULT_SPSPOP_PARAMS,
-        n_key,
-        lc_halopop_z_obs_multi_z,
-        lc_halopop_t_obs_multi_z,
-        lc_halopop_mah_params_multi_z,
-        lc_halopop_logmp0_multi_z,
-        lc_halopop_nhalos_multi_z,
-        lc_halopop_vol_mpc3_multi_z,
-        t_table,
-        ssp_data,
-        precomputed_ssp_mag_table_multi_z,
-        z_phot_table_multi_z,
-        wave_eff_table_multi_z,
-        DEFAULT_MZR_PARAMS,
-        DEFAULT_SCATTER_PARAMS,
-        ZERO_SSPERR_PARAMS,
-        lh_centroids_multi_z,
-        dmag_centroids_multi_z,
-        mag_columns,
-        mag_thresh_column,
-        mag_thresh,
-        DEFAULT_COSMOLOGY,
-        FB,
-        frac_cat,
-    )
-
-    lg_n_multi_z, lg_n_avg_err_multi_z = n_mag.n_mag_kern_multi_z(
-        DIFFSTARPOP_UM_plus_exsitu, *n_args_multi_z
-    )
-    assert np.isfinite(lg_n_multi_z).all()
-    assert np.isfinite(lg_n_avg_err_multi_z).all()
-
-    for zbin in range(0, len(zbins)):
-        zmin = zbins[zbin][0]
-        zmax = zbins[zbin][1]
-
-        z_phot_table = jnp.linspace(zmin, zmax, n_z_phot_table)
-
-        precomputed_ssp_mag_table = psspp.get_precompute_ssp_mag_redshift_table(
-            tcurves, ssp_data, z_phot_table, DEFAULT_COSMOLOGY
-        )
-
-        wave_eff_table = lc_phot_kern.get_wave_eff_table(z_phot_table, tcurves)
-
-        lc_halopop = lc_halopop_multi_z[zbin]
-
-        lh_centroids = jnp.asarray(
-            np.load(
-                os.path.join(
-                    LH_CENTROIDS_PATH,
-                    "feniks_lh_centroids_z_"
-                    + str(zbins[zbin][0])
-                    + "-"
-                    + str(zbins[zbin][1])
-                    + "_test.npy",
-                )
-            )
-        )
-        dmag_centroids = jnp.ones((lh_centroids.shape[0], 1)) * dmag
-
-        n_args_single_z = (
-            DIFFSTARPOP_UM_plus_exsitu,
-            DEFAULT_SPSPOP_PARAMS,
-            n_key,
-            jnp.array(lc_halopop["z_obs"]),
-            lc_halopop["t_obs"],
-            lc_halopop["mah_params"],
-            lc_halopop["logmp0"],
-            lc_halopop["nhalos"],
-            lc_halopop["lc_vol_Mpc3"],
-            t_table,
-            ssp_data,
-            precomputed_ssp_mag_table,
-            z_phot_table,
-            wave_eff_table,
-            DEFAULT_MZR_PARAMS,
-            DEFAULT_SCATTER_PARAMS,
-            ZERO_SSPERR_PARAMS,
-            lh_centroids,
-            dmag_centroids,
-            mag_columns,
-            mag_thresh_column,
-            mag_thresh,
-            DEFAULT_COSMOLOGY,
-            FB,
-        )
-        lg_n_single_z, lg_n_avg_err_single_z = n_mag.n_mag_kern(*n_args_single_z)
-
-        assert np.isfinite(lg_n_single_z).all()
-        assert np.isfinite(lg_n_avg_err_single_z).all()
-        assert np.allclose(lg_n_multi_z[zbin], lg_n_single_z)
